+++ LICENSE
@@ -246,7 +246,7 @@ in one of these ways:
     (including a physical distribution medium), accompanied by a
     written offer, valid for at least three years and valid for as
     long as you offer spare parts or customer support for that product
-    tileBasedKendallAlgorithms, to give anyone who possesses the object code either (1) a
+    sparkBasedKendallAlgorithms, to give anyone who possesses the object code either (1) a
     copy of the Corresponding Source for all the software in the
     product that is covered by this License, on a durable physical
     medium customarily used for software interchange, for a price no
