--- src/main/java/listBasedKendallAlgorithms/listBasedTiles/tiles/DoublePair.java
@@ -1,19 +1,0 @@
-package listBasedKendallAlgorithms.listBasedTiles.tiles;
-
-public class DoublePair {
-    private final double x;
-    private final double y;
-
-    public DoublePair(double x, double y) {
-        this.x = x;
-        this.y = y;
-    }
-
-    public double getX() {
-        return x;
-    }
-
-    public double getY() {
-        return y;
-    }
-}
