+++ src/main/java/sparkBasedKendallAlgorithms/reader/DatasetReader.java
@@ -1,4 +1,4 @@
-package tileBasedKendallAlgorithms.reader;
+package sparkBasedKendallAlgorithms.reader;
 
 import org.apache.spark.sql.Dataset;
 import org.apache.spark.sql.Row;
