+++ src/test/java/listBasedKendallAlgorithms/listBasedTiles/ListBasedTileBasedKendallManagerTest.java
@@ -5,6 +5,7 @@ import static org.junit.Assert.assertEquals;
 
 import java.io.IOException;
 
+import listBasedKendallAlgorithms.ListBasedTileBasedKendallManager;
 import listBasedKendallAlgorithms.listBasedReader.ColumnPair;
 import listBasedKendallAlgorithms.listBasedReader.Reader;
