+++ src/test/java/tileBasedkendallAlgorithms/sparkSetup/SparkSetupTest.java
@@ -2,7 +2,8 @@ package tileBasedkendallAlgorithms.sparkSetup;
 
 import org.apache.spark.sql.SparkSession;
 import org.junit.Test;
-import tileBasedKendallAlgorithms.sparkSetup.SparkSetup;
+
+import sparkBasedKendallAlgorithms.sparkSetup.SparkSetup;
 
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertNotNull;
